# Acceleration of fitting by avoiding model call overhead