As described in #100, the functions that often call DeerLab's built-in models suffer from some heavy overhead leading to a serious slow-down of the runtime.  

By implementing the way of avoiding call overhead described in #100 we can speed up the execution of the fit functions. Mostly affected are: 
- **fitsignal**: about a factor x2 acceleration
- **fitmultimodel**: about a factor x7-14 acceleration

Here are some benchmark results showing the acceleration in more detail: 

#### ``fitsignal`` (before)
```
=================================================== test session starts 
platform win32 -- Python 3.8.0, pytest-6.2.2, py-1.10.0, pluggy-0.13.1
rootdir: D:\lufa\projects\DeerLab\DeerLab
collected 44 items                                                                                                                                                                                                                                                                                                  

test\test_fitsignal.py ............................................                                                                                                                                                                                                                                          [100%] 

================================================== slowest 10 durations 
24.30s call     test/test_fitsignal.py::test_ridme5
11.39s call     test/test_fitsignal.py::test_cost_value
10.31s call     test/test_fitsignal.py::test_ovl4pdeer
9.51s call     test/test_fitsignal.py::test_ridme3
9.00s call     test/test_fitsignal.py::test_5pdeer
6.87s call     test/test_fitsignal.py::test_7pdeer
6.12s call     test/test_fitsignal.py::test_global_4pdeer
4.63s call     test/test_fitsignal.py::test_global_full_parametric
4.31s call     test/test_fitsignal.py::test_global_scale_4pdeer
3.83s call     test/test_fitsignal.py::test_boundaries_adjust_ex
================================================== 44 passed in 133.21s
```
#### ``fitsignal`` (now)
```
================================================== test session starts 
rootdir: D:\lufa\projects\DeerLab\DeerLab
collected 44 items                                                                                                                                                                                                                                                                                                  

test\test_fitsignal.py ............................................                                                                                                                                                                                                                                          [100%] 

================================================== slowest 10 durations 
6.15s call     test/test_fitsignal.py::test_ovl4pdeer
5.43s call     test/test_fitsignal.py::test_7pdeer
3.95s call     test/test_fitsignal.py::test_cost_value
3.78s call     test/test_fitsignal.py::test_ridme5
3.35s call     test/test_fitsignal.py::test_5pdeer
2.89s call     test/test_fitsignal.py::test_boundaries_adjust_ex
2.31s call     test/test_fitsignal.py::test_ridme1
2.07s call     test/test_fitsignal.py::test_ridme3
1.92s call     test/test_fitsignal.py::test_global_4pdeer
1.90s call     test/test_fitsignal.py::test_plot
================================================== 44 passed in 54.60s
```
#### ``fitmultimodel`` (before)
```
====================================================================================== test session starts 
platform win32 -- Python 3.8.0, pytest-6.2.2, py-1.10.0, pluggy-0.13.1
rootdir: D:\lufa\projects\DeerLab\DeerLab
collected 20 items                                                                                                                                                                                                                                                                                                  

test\test_fitmultimodel.py ....................                                                                                                                                                                                                                                                              [100%] 

====================================================================================== slowest 10 durations  
15.87s call     test/test_fitmultimodel.py::test_rescaling
14.08s call     test/test_fitmultimodel.py::test_multirice_split
13.98s call     test/test_fitmultimodel.py::test_multirice
9.75s call     test/test_fitmultimodel.py::test_global_multirice
8.60s call     test/test_fitmultimodel.py::test_multigauss_spread
7.23s call     test/test_fitmultimodel.py::test_multirice_spread
6.59s call     test/test_fitmultimodel.py::test_multigauss_merge
6.54s call     test/test_fitmultimodel.py::test_multirice_merge
5.23s call     test/test_fitmultimodel.py::test_multigauss
5.08s call     test/test_fitmultimodel.py::test_plot
======================================================================================20 passed in 128.47s (0:02:08)
```
#### ``fitmultimodel`` (now)
```
====================================================================================== test session starts 
platform win32 -- Python 3.8.0, pytest-6.2.2, py-1.10.0, pluggy-0.13.1
rootdir: D:\lufa\projects\DeerLab\DeerLab
collected 20 items                                                                                                                                                                                                                                                                                                  

test\test_fitmultimodel.py ....................                                                                                                                                                                                                                                                              [100%] 

====================================================================================== slowest 10 durations  
2.05s call     test/test_fitmultimodel.py::test_background_fit
0.75s call     test/test_fitmultimodel.py::test_plot
0.69s call     test/test_fitmultimodel.py::test_multirice
0.69s call     test/test_fitmultimodel.py::test_multirice_split
0.45s call     test/test_fitmultimodel.py::test_rescaling
0.44s call     test/test_fitmultimodel.py::test_multirice_spread
0.42s call     test/test_fitmultimodel.py::test_multigauss
0.38s call     test/test_fitmultimodel.py::test_multirice_merge
0.33s call     test/test_fitmultimodel.py::test_global_multirice
0.31s call     test/test_fitmultimodel.py::test_multigauss_spread
====================================================================================== 20 passed in 8.90s
```

Closes #100